change doubles to ints

+++ Emulator/src/AssemblyEmulator.java
@@ -5,13 +5,13 @@ import java.util.*;
 
 public class AssemblyEmulator {
 
-    private Map<String, Double> registers;
+    private Map<String, Integer> registers;
     private Map<String, Integer> functions;
     private Map<Integer, String> returns;
     private Map<String, Integer> functionCalls;
     private ArrayList<String> list;
-    private double[] memory;
-    private double rv;
+    private int[] memory;
+    private int rv;
 
     public AssemblyEmulator(FileReader file) throws IOException {
         init();
@@ -34,6 +34,8 @@ public class AssemblyEmulator {
         printMemory();
         System.out.println("RV: " + rv);
     }
+
+
 
     private void fillReturnsIndex(){
         for(int i = 0 ; i < list.size(); i++){
@@ -60,7 +62,7 @@ public class AssemblyEmulator {
         registers = new HashMap<>();
         functions = new HashMap<>();
         functionCalls = new HashMap<>();
-        memory = new double[1]; //saved pc
+        memory = new int[1]; //saved pc
     }
 
     private void printMemory() {
@@ -78,7 +80,7 @@ public class AssemblyEmulator {
             } else if (left.startsWith("S")) {
                 allocateMemory(line);
             } else if (line.startsWith("M")) {
-                fillMemoryArray(Double.parseDouble(left), line);
+                fillMemoryArray(Integer.parseInt(left), line);
             } else if (line.startsWith("B")) {
                 return branches(line, numberOfLine);
             } else if (line.startsWith("J")) {
@@ -98,7 +100,7 @@ public class AssemblyEmulator {
                     callFunction(str);
                     resizeMemory(memory.length - 1); //delete saved pc
                 } else {
-                    Double result = getAddress(line.substring(4, line.length() - 1));
+                    int result = getAddress(line.substring(4, line.length() - 1));
                     //memory[(int)result]
                 }
             } else if(line.startsWith("RET")){
@@ -143,13 +145,13 @@ public class AssemblyEmulator {
     private int branches(String str, int numberOfLine) {
         String type = str.substring(0, 3);
         int index = str.indexOf(',');
-        double a = getValue(str.substring(3, index), numberOfLine);
-        double b = getValue(str.substring(index + 1, str.indexOf(',', index + 1)), numberOfLine);
+        int a = getValue(str.substring(3, index), numberOfLine);
+        int b = getValue(str.substring(index + 1, str.indexOf(',', index + 1)), numberOfLine);
         int pc = (int) getValue(str.substring(str.indexOf(',', index + 1) + 1, str.indexOf(";")), numberOfLine);
         return compareValues(type, a, b, pc);
     }
 
-    private int compareValues(String type, double n, double m, int pc) {
+    private int compareValues(String type, int n, int m, int pc) {
         boolean isTrue = false;
         if (type.equals("BLT")) {
             if (n < m) isTrue = true;
@@ -168,7 +170,7 @@ public class AssemblyEmulator {
         return -1;
     }
 
-    private double getValue(String str, int numberOfLine) {
+    private int getValue(String str, int numberOfLine) {
         //M[sp] ეგეთი ოპერაცია არ მოსულაო, ეწერა.
         if(str.startsWith("RV")){
             return rv;
@@ -176,7 +178,7 @@ public class AssemblyEmulator {
         if (str.startsWith("R") && !str.startsWith("RET") && !str.startsWith("RV")) {
             return registers.get(str);
         } else if (isNumber(str)) {
-            return Double.parseDouble(str);
+            return Integer.parseInt(str);
         } else {
             if (containsOperator(str)) {
                 char operator = getOperator(str);
@@ -192,18 +194,18 @@ public class AssemblyEmulator {
         }
     }
 
-    private void fillMemoryArray(double location, String line) {
+    private void fillMemoryArray(int location, String line) {
         if (location >= memory.length) {
             System.out.println(location + " " + memory.length);
             System.out.println("error");
         } else {
             String str = line.substring(line.indexOf("=") + 1, line.indexOf(";"));
-            double res = computeBytes(str);
-            memory[(int) location] = res;
+            int res = computeBytes(str);
+            memory[location] = res;
         }
     }
 
-    private double computeBytes(String str) {
+    private int computeBytes(String str) {
         char c = '0';
         boolean toInt = false;
         if (str.startsWith(".")) {
@@ -213,7 +215,7 @@ public class AssemblyEmulator {
             toInt = true;
             str = str.substring(4);
         }
-        double res = getValueOfTheRightSide(str);
+        int res = getValueOfTheRightSide(str);
         if (toInt) {
             int number = (int) res;
             res = number;
@@ -258,18 +260,18 @@ public class AssemblyEmulator {
         int index = line.indexOf("=");
         String leftSide = line.substring(0, index);
         String str = line.substring(index + 1, line.indexOf(";"));
-        double rightSide = computeBytes(str);
+        int rightSide = computeBytes(str);
         registers.put(leftSide, rightSide);
         System.out.println(leftSide + " " + rightSide + " registers");
     }
 
-    private double getAddress(String str) {
+    private int getAddress(String str) {
         if (isNumber(str)) {
-            return Double.parseDouble(str) / 4; //zustad ar vici es
+            return Integer.parseInt(str) / 4; //zustad ar vici es
         } else if (containsOperator(str)) {
             //an sp + a an a + sp an a + r an r + 1 an r1 + r2
             if (str.contains("SP")) {
-                double res = computeResult(str);
+                int res = computeResult(str);
                 return res;
             }
             return computeResult(str) / 4;
@@ -282,17 +284,17 @@ public class AssemblyEmulator {
         }
     }
 
-    private double getValueOfTheRightSide(String str) {
+    private int getValueOfTheRightSide(String str) {
         if (isNumber(str)) { //loads constants.
-            return Double.parseDouble(str);
+            return Integer.parseInt(str);
         } else if (str.charAt(0) == '-' && isNumber(str.substring(1))) { //is negative number.
-            return -1 * Double.parseDouble(str.substring(1));
+            return -1 * Integer.parseInt(str.substring(1));
         } else if (containsOperator(str) && !str.startsWith("M")) {
             //an sp + a an a + sp an a + r an r + 1 an r1 + r2
             return computeResult(str);
         } else if (str.startsWith("M")) {
-            double res = getAddress(str.substring(str.indexOf('[') + 1, str.indexOf(']')));
-            return memory[(int) res];
+            int res = getAddress(str.substring(str.indexOf('[') + 1, str.indexOf(']')));
+            return memory[res];
         } else if(str.startsWith("RV")) {
             return rv;
         } else {
@@ -304,16 +306,16 @@ public class AssemblyEmulator {
         }
     }
 
-    private double computeResult(String str) {
+    private int computeResult(String str) {
         char operator = getOperator(str);
         int index = str.indexOf(operator);
         String first = str.substring(0, index);
         String second = str.substring(index + 1);
         if(first.startsWith("RV")){
-            first = Double.toString(rv);
+            first = Integer.toString(rv);
         }
         if(second.startsWith("RV")){
-            second = Double.toString(rv);
+            second = Integer.toString(rv);
         }
         if (isNumber(first)) {
             return computeValue(first, second, operator, true);
@@ -322,22 +324,22 @@ public class AssemblyEmulator {
         } else if (isNumber(first) && isNumber(second)) {
             return Integer.parseInt(first) + Integer.parseInt(second);
         } else {
-            double value1 = registers.get(first);
-            double value2 = registers.get(second);
+            int value1 = registers.get(first);
+            int value2 = registers.get(second);
             return compute(value1, value2, operator);
         }
     }
 
-    private double computeValue(String first, String second, char operator, boolean isOrdered) {
+    private int computeValue(String first, String second, char operator, boolean isOrdered) {
         if (second.startsWith("S")) {
             return findMemoryArrayIndex(operator, Integer.parseInt(first));
         } else {
             //starts with R.
-            double val = registers.get(second);
+            int val = registers.get(second);
             if (isOrdered) {
-                return compute(Double.parseDouble(first), val, operator);
+                return compute(Integer.parseInt(first), val, operator);
             } else {
-                return compute(val, Double.parseDouble(first), operator);
+                return compute(val, Integer.parseInt(first), operator);
             }
         }
     }
@@ -357,7 +359,7 @@ public class AssemblyEmulator {
     }
 
     private void deleteSavedPC(){
-        double[] curr = new double[memory.length - 1];
+        int[] curr = new int[memory.length - 1];
         for (int i = 1; i < memory.length; i++) {
             curr[i - 1] = memory[i];
         }
@@ -366,7 +368,7 @@ public class AssemblyEmulator {
 
     private void resizeMemory(int size) {
        // System.out.println("size " + size);
-        double[] curr = new double[size];
+        int[] curr = new int[size];
         for (int i = 0; i < Math.min(size, memory.length); i++) {
             curr[i] = memory[i];
         }
@@ -390,8 +392,8 @@ public class AssemblyEmulator {
             return str.substring(0, str.indexOf("="));
         } else if (str.startsWith("M")) {
             String address = str.substring(str.indexOf("[") + 1, str.indexOf("]"));
-            double res = getAddress(address);
-            return Double.toString(res);
+            int res = getAddress(address);
+            return Integer.toString(res);
         } else if (str.startsWith("S")) {
             return str.substring(0, str.indexOf("="));
         }
@@ -402,7 +404,7 @@ public class AssemblyEmulator {
         return address.indexOf('+') != -1 || address.indexOf('-') != -1 || address.indexOf('*') != -1 || address.indexOf('/') != -1;
     }
 
-    private double compute(double a, double b, char operator) {
+    private int compute(int a, int b, char operator) {
         if (operator == '+') {
             return a + b;
         } else if (operator == '-') {
